moved all controller files to toplevel and fixed some more uses
