+++ okapi/lib/OCPLAccessLogs.php
@@ -3,8 +3,8 @@
 namespace okapi\lib;
 
 use okapi\Settings;
-use okapi\Db;
-use okapi\Request\OkapiRequest;
+use okapi\core\Db;
+use okapi\core\Request\OkapiRequest;
 
 /**
  * This is a (hopefully temporary) class which holds all functionality related
